+++ src/main/java/com/shinhan/knockknock/domain/dto/cardhistory/ReadCardHistoryResponse.java
@@ -1,4 +1,4 @@
-package com.shinhan.knockknock.domain.dto;
+package com.shinhan.knockknock.domain.dto.cardhistory;
 
 import com.shinhan.knockknock.domain.entity.CardCategoryEntity;
 import com.shinhan.knockknock.domain.entity.CardEntity;
