+++ src/main/java/com/shinhan/knockknock/domain/dto/welfarebook/ReadWelfareBookResponse.java
@@ -1,4 +1,4 @@
-package com.shinhan.knockknock.domain.dto;
+package com.shinhan.knockknock.domain.dto.welfarebook;
 
 import com.shinhan.knockknock.domain.entity.UserEntity;
 import com.shinhan.knockknock.domain.entity.WelfareEntity;
@@ -14,9 +14,7 @@ import java.sql.Timestamp;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class CreateWelfareBookRequest {
-    @Schema(example = "1")
-    private Long welfareBookNo;
+public class ReadWelfareBookResponse {
     @Schema(example = "2024-08-08 09:00")
     private Timestamp welfareBookStartDate;
     @Schema(example = "2024-08-08 13:00")
