+++ src/main/java/com/shinhan/knockknock/service/WelfareServiceImpl.java
@@ -1,13 +1,12 @@
 package com.shinhan.knockknock.service;
 
-import com.shinhan.knockknock.domain.dto.CreateWelfareRequest;
-import com.shinhan.knockknock.domain.dto.ReadWelfareResponse;
+import com.shinhan.knockknock.domain.dto.welfare.CreateWelfareRequest;
+import com.shinhan.knockknock.domain.dto.welfare.ReadWelfareResponse;
 import com.shinhan.knockknock.domain.entity.WelfareEntity;
 import com.shinhan.knockknock.repository.WelfareRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.function.Function;
